Ruleset utils + testfixes.
