+++ src/OmniSharp.MSBuild/AssemblyInfo.cs
@@ -2,3 +2,4 @@
 
 [assembly: InternalsVisibleTo("OmniSharp.MSBuild.Tests")]
 [assembly: InternalsVisibleTo("TestUtility")]
+[assembly: InternalsVisibleTo("OmniSharp.Roslyn.CSharp.Tests")]
