+++ tests/OmniSharp.Roslyn.CSharp.Tests/CustomRoslynAnalyzerFacts.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
 using OmniSharp.Models.Diagnostics;
+using OmniSharp.MSBuild;
 using OmniSharp.Roslyn.CSharp.Services.Diagnostics;
 using TestUtility;
 using Xunit;
@@ -146,17 +147,13 @@ namespace OmniSharp.Roslyn.CSharp.Tests
                             Console.WriteLine(null); // This is CS0162, unreachable code.
                         }
                     }");
-                // var ruleService = host.GetExport<RulesetsForProjects>();
 
-                // var projectIds = AddProjectWitFile(host, testFile);
+                var projectId = AddProjectWitFile(host, testFile);
+                var testRules = CreateRules("CS0162", ReportDiagnostic.Hidden);
 
-                // var testRules = CreateRules("CS0162", ReportDiagnostic.Hidden);
+                var project = host.Workspace.CurrentSolution.GetProject(projectId);
 
-                // ruleService.AddOrUpdateRuleset(projectIds.Single(), new RuleSet(
-                //     "",
-                //     new ReportDiagnostic(),
-                //     testRules.ToImmutableDictionary(),
-                //     new ImmutableArray<RuleSetInclude>()));
+                RulesetUtils.UpdateProjectWithRulesets(host.Workspace, project, testRules.ToImmutableDictionary());
 
                 var result = await host.RequestCodeCheckAsync();
 
@@ -170,18 +167,15 @@ namespace OmniSharp.Roslyn.CSharp.Tests
             using (var host = GetHost())
             {
                 var testFile = new TestFile("testFile_2.cs", "class _this_is_invalid_test_class_name { int n = true; }");
-                // var ruleService = host.GetExport<RulesetsForProjects>();
 
                 var testAnalyzerRef = new TestAnalyzerReference("TS1100");
 
-                var projectIds = AddProjectWitFile(host, testFile, testAnalyzerRef);
+                var projectId = AddProjectWitFile(host, testFile, testAnalyzerRef);
                 var testRules = CreateRules(testAnalyzerRef.Id.ToString(), ReportDiagnostic.Hidden);
 
-                // ruleService.AddOrUpdateRuleset(projectIds.Single(), new RuleSet(
-                //     "",
-                //     new ReportDiagnostic(),
-                //     testRules.ToImmutableDictionary(),
-                //     new ImmutableArray<RuleSetInclude>()));
+                var project = host.Workspace.CurrentSolution.GetProject(projectId);
+
+                RulesetUtils.UpdateProjectWithRulesets(host.Workspace, project, testRules.ToImmutableDictionary());
 
                 var result = await host.RequestCodeCheckAsync("testFile_2.cs");
                 Assert.Contains(result.QuickFixes.OfType<DiagnosticLocation>(), f => f.Text.Contains(testAnalyzerRef.Id.ToString()) && f.LogLevel == "Hidden");
@@ -204,19 +198,15 @@ namespace OmniSharp.Roslyn.CSharp.Tests
             {
                 var testFile = new TestFile("testFile_3.cs", "class _this_is_invalid_test_class_name { int n = true; }");
 
-                // var ruleService = host.GetExport<RulesetsForProjects>();
-
                 var testAnalyzerRef = new TestAnalyzerReference("TS1101");
 
-                var projectIds = AddProjectWitFile(host, testFile, testAnalyzerRef);
+                var projectId = AddProjectWitFile(host, testFile, testAnalyzerRef);
 
                 var testRules = CreateRules(testAnalyzerRef.Id.ToString(), ReportDiagnostic.Suppress);
 
-                // ruleService.AddOrUpdateRuleset(projectIds.Single(), new RuleSet(
-                //     "",
-                //     new ReportDiagnostic(),
-                //     testRules.ToImmutableDictionary(),
-                //     new ImmutableArray<RuleSetInclude>()));
+                var project = host.Workspace.CurrentSolution.GetProject(projectId);
+
+                RulesetUtils.UpdateProjectWithRulesets(host.Workspace, project, testRules.ToImmutableDictionary());
 
                 var result = await host.RequestCodeCheckAsync("testFile_3.cs");
                 Assert.DoesNotContain(result.QuickFixes, f => f.Text.Contains(testAnalyzerRef.Id.ToString()));
@@ -248,17 +238,18 @@ namespace OmniSharp.Roslyn.CSharp.Tests
             }
         }
 
-        private IEnumerable<ProjectId> AddProjectWitFile(OmniSharpTestHost host, TestFile testFile, TestAnalyzerReference testAnalyzerRef = null)
+        private ProjectId AddProjectWitFile(OmniSharpTestHost host, TestFile testFile, TestAnalyzerReference testAnalyzerRef = null)
         {
             var analyzerReferences = testAnalyzerRef == null ? default :
                 new AnalyzerReference[] { testAnalyzerRef }.ToImmutableArray();
 
-            return TestHelpers.AddProjectToWorkspace(
+            return  TestHelpers.AddProjectToWorkspace(
                             host.Workspace,
                             "project.csproj",
                             new[] { "netcoreapp2.1" },
                             new[] { testFile },
-                            analyzerRefs: analyzerReferences);
+                            analyzerRefs: analyzerReferences)
+                    .Single();
         }
     }
 }
